// src/app/openapi/model/coin.ts
/**
 * alfa
 * beta
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: 48928310+averge@users.noreply.github.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class Coin { 
    id?: string;
    nombre?: string;
    sufijo?: string;
    precio?: number;
    descripcion?: string;
    link?: string;

    constructor(nombre?: string, sufijo?: string, precio?: number, descripcion?: string, link?: string) {
        this.nombre = nombre;
        this.sufijo = sufijo;
        this.precio = precio;
        this.descripcion = descripcion;
        this.link = link;
    }
}